GtkSelectionData length is off by one https://bugs.webkit.org/show_bug.cgi?id=113962
Reviewed by Anders Carlsson.

No new tests. Since drag data is interpreted as a null-terminated string
this is difficult to test with a C program and we have no infrastructure
built for using GIR tests.

* platform/gtk/PasteboardHelper.cpp:
(WebCore::PasteboardHelper::fillSelectionData): Instead of including the null
character in the paste data length, just include the string. This matches the behavior
of Firefox.

Identifier: 134768@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@150350 268f45cc-cd09-0410-ab3c-d52691b4dbfc