Remove ChromeClient::webView() https://bugs.webkit.org/show_bug.cgi?id=116054
Source/WebCore:

Reviewed by Darin Adler.

This blatantly horrible layer violation was only used to know if a ChromeClient is an empty
client or not. We already have a (slightly less horrible) way to do that.

* loader/EmptyClients.h:
* page/ChromeClient.h:
(ChromeClient):

Source/WebKit/blackberry:

Reviewed by Andreas Kling.

This blatantly horrible layer violation was only used by the Chromium port; get rid of it.

* WebCoreSupport/ChromeClientBlackBerry.h:
(ChromeClientBlackBerry):

Source/WebKit/efl:

Reviewed by Darin Adler.

This blatantly horrible layer violation was only used to know if a ChromeClient is an empty
client or not. We already have a (slightly less horrible) way to do that.

* WebCoreSupport/ChromeClientEfl.h:
(ChromeClientEfl):

Source/WebKit/gtk:

Reviewed by Darin Adler.

This blatantly horrible layer violation was only used to know if a ChromeClient is an empty
client or not. We already have a (slightly less horrible) way to do that.

* WebCoreSupport/ChromeClientGtk.h:
(ChromeClient):

Source/WebKit/mac:

Reviewed by Darin Adler.

This blatantly horrible layer violation was only used to know if a ChromeClient is an empty
client or not. We already have a (slightly less horrible) way to do that.

* WebCoreSupport/WebChromeClient.h:
(WebChromeClient):

Source/WebKit/qt:

Reviewed by Darin Adler.

This blatantly horrible layer violation was only used to know if a ChromeClient is an empty
client or not. We already have a (slightly less horrible) way to do that.

* WebCoreSupport/ChromeClientQt.h:
(ChromeClientQt):

Source/WebKit/win:

Reviewed by Darin Adler.

This blatantly horrible layer violation was only used to know if a ChromeClient is an empty
client or not. We already have a (slightly less horrible) way to do that.

* WebCoreSupport/WebChromeClient.h:
(WebChromeClient):

Source/WebKit/wince:

Reviewed by Darin Adler.

This blatantly horrible layer violation was only used to know if a ChromeClient is an empty
client or not. We already have a (slightly less horrible) way to do that.

* WebCoreSupport/ChromeClientWinCE.h:
(ChromeClientWinCE):

Identifier: 134771@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@150353 268f45cc-cd09-0410-ab3c-d52691b4dbfc