PLATFORM(*) macros used in Source/WebCore/loader/archive/ArchiveFactory.cpp https://bugs.webkit.org/show_bug.cgi?id=116453
Reviewed by Alexey Proskuryakov.

* loader/archive/ArchiveFactory.cpp: Remove a redundant PLATFORM(QT) macro when checking whether
to include the LegacyWebArchive.h header that's specific to the WebArchive support. The Qt port
does not enable the WebArchive feature at all (which is enabled only on AppleMac, AppleWin and iOS
platforms) so there's no reason to additionally condition the header include with building on
non-Qt platforms.


Identifier: 134851@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@150444 268f45cc-cd09-0410-ab3c-d52691b4dbfc