Fix assertion in the getComputedStyle-background-shorthand.html test https://bugs.webkit.org/show_bug.cgi?id=116639
Reviewed by Ryosuke Niwa.

getComputedStyle-background-shorthand.html was asserting under the newly added
code in willBeRemovedFromTree(). The assertion was caused by too many calls to
FrameView::removeSlowRepaintObject(), which happened because we called it for
RenderTexts, which share style with their parent renderer.

RenderText overrides styleWillChange() to do nothing, so we never
call add/removeSlowRepaintObject() for them. Thus we should also skip this
work in willBeRemovedFromTree() for RenderTexts.

* rendering/RenderObject.cpp:
(WebCore::RenderObject::willBeRemovedFromTree):

Identifier: 134940@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@150547 268f45cc-cd09-0410-ab3c-d52691b4dbfc