Don't create an accelerated compositing layer unless we have to for snapshotting plug-ins https://bugs.webkit.org/show_bug.cgi?id=117076 <rdar://problem/14030364>
Reviewed by Beth Dakin and Anders Carlsson.

If a plug-in can support software snapshotting, then don't create an
accelerated compositing layer for it while in snapshot mode. This is a
better user experience -- otherwise the page/plug-in looks like it broke.

* WebProcess/Plugins/PluginView.cpp:
(WebKit::PluginView::isAcceleratedCompositingEnabled): If we are in snapshotting mode
and we are a Flash plugin (the only one that supports software snapshots at the moment)
then do not create an accelerated compositing layer.

Identifier: 135344@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151026 268f45cc-cd09-0410-ab3c-d52691b4dbfc