Scrolling with platformWidget and delegateScrolling is incorrectly clamped https://bugs.webkit.org/show_bug.cgi?id=117369 <rdar://problem/13985064>
Patch by Benjamin Poulain <bpoulain@apple.com> on 2013-06-09
Reviewed by Darin Adler.

The patch http://trac.webkit.org/changeset/142526 clamps the input position
in the ScrollView scroll range. This was done for the path to the ScrollingCoordinator.

The problem with that change is ScrollView::setScrollPosition() can delegate the scrolling
to either a platformWidget, or through delegateScrolling. After r142526, the position is clamped,
and we do not let those external scrolling mechanims handle the out-of-bound scrolling.

This patch fixes the issue by moving the threaded scrolling call to ScrollView,
after the delegate handling code.

* page/FrameView.cpp:
(WebCore::FrameView::setScrollPosition):
(WebCore::FrameView::requestScrollPositionUpdate):


Identifier: 135612@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151356 268f45cc-cd09-0410-ab3c-d52691b4dbfc