[CSS Regions] Regions don't create a stacking context for their contents https://bugs.webkit.org/show_bug.cgi?id=114268
Reviewed by Alexandru Chiculita.

Source/WebCore:

Paint the region content (from RenderFlowThread) during the background phase instead of the
foreground phase.

Let's assume there is a region and a sibling element. The sibling partially overlaps the
region and appears after the region in DOM order.

Before this patch, we would paint in this order:
1) Region background
2) Sibling background
3) Region content
4) Sibling content

As a result, the sibling background would incorrectly paint between the region
background and the region content. Instead, the sibling background should paint over the
region background and the region content.

After this patch, we paint in this order:
1) Region background
2) Region content
3) Sibling background
4) Sibling content

Test: fast/regions/region-sibling-paint-order.html

* rendering/RenderRegion.cpp:
(WebCore::RenderRegion::paintObject):

LayoutTests:

* fast/regions/region-sibling-paint-order-expected.html: Added.
* fast/regions/region-sibling-paint-order.html: Added.


Identifier: 135686@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151475 268f45cc-cd09-0410-ab3c-d52691b4dbfc