Source/WebCore: Whitespace between inlines with nowrap and a shrink-to-fit parent gets a line-break when it shouldn't https://bugs.webkit.org/show_bug.cgi?id=117370
Reviewed by David Hyatt.

Test: fast/text/whitespace/inline-whitespace-wrapping-8.html

A no-wrap inline shouldn't include trailing space when deciding whether it fits on a line.
Likewise when we finish iterating through the objects on a line we should clear our linebreak
if the only thing that prevents us fitting on the line is our collapsed trailing whitespace.

Removing the trailing space from this measurement means we need to watch out for potential
breaks between no-wrap inlines, in particular if we leave a no-wrap inline ignoring spaces
and enter an autowrap inline then we need to mark the beginning of the autowrap inline
as a potential linebreak. The test fast/text/whitespace/inline-whitespace-wrapping-5.html
is an example of such a case.

* rendering/RenderBlockLineLayout.cpp:
(WebCore::LineWidth::LineWidth):
(WebCore::LineWidth::fitsOnLine):
(WebCore::LineWidth::fitsOnLineExcludingTrailingWhitespace):
(WebCore::LineWidth::fitsOnLineExcludingTrailingCollapsedWhitespace):
(WebCore::LineWidth::setTrailingWhitespaceWidth):
(WebCore::RenderBlock::LineBreaker::nextSegmentBreak):

LayoutTests: Whitespace between inlines with nowrap and a shrink-to-fit parent line-break incorrectly
https://bugs.webkit.org/show_bug.cgi?id=117370

Reviewed by David Hyatt.

* fast/text/whitespace/inline-whitespace-wrapping-8-expected.html: Added.
* fast/text/whitespace/inline-whitespace-wrapping-8.html: Added.


Identifier: 135724@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151518 268f45cc-cd09-0410-ab3c-d52691b4dbfc